Split line_input/mod.rs and tree/mod.rs to stay under 1000 lines
Extracts the update() logic from line_input into update.rs, and
extracts rendering helpers from tree into render.rs, matching the
pattern used by chart, conversation_view, heatmap, and other
components with separate render modules.
